+++ src/main/kotlin/com/project/indistraw/domain/account/adapter/input/AccountWebAdapter.kt
@@ -4,9 +4,9 @@ import com.project.indistraw.domain.account.adapter.input.data.request.UpdateAcc
 import com.project.indistraw.domain.account.adapter.input.data.request.UpdateAddressRequest
 import com.project.indistraw.domain.account.adapter.input.data.request.UpdatePasswordRequest
 import com.project.indistraw.domain.account.adapter.input.data.response.AccountInfoResponse
-import com.project.indistraw.domain.account.adapter.input.data.response.FilmographyResponse
 import com.project.indistraw.domain.account.adapter.input.mapper.AccountDataMapper
 import com.project.indistraw.domain.account.application.port.input.*
+import com.project.indistraw.domain.movie.application.port.input.MyFilmographyUseCase
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -23,8 +23,7 @@ class AccountWebAdapter(
     private val findAccountInfoUseCase: FindAccountInfoUseCase,
     private val accountWithdrawUseCase: AccountWithdrawUseCase,
     private val mappingActorUseCase: MappingActorUseCase,
-    private val mappingDirectorUseCase: MappingDirectorUseCase,
-    private val myFilmographyUseCase: MyFilmographyUseCase
+    private val mappingDirectorUseCase: MappingDirectorUseCase
 ) {
 
     @GetMapping("/phone-number/{phoneNumber}")
@@ -72,11 +71,5 @@ class AccountWebAdapter(
     fun mappingDirector(@PathVariable idx: Long): ResponseEntity<Void> =
         mappingDirectorUseCase.execute(idx)
             .run { ResponseEntity.status(HttpStatus.CREATED).build() }
-
-    @GetMapping("/filmography")
-    fun findFilmography(): ResponseEntity<List<FilmographyResponse>> =
-        myFilmographyUseCase.execute()
-            .map { accountDataMapper toResponse it }
-            .let { ResponseEntity.ok(it) }
 
 }
