--- src/main/kotlin/com/project/indistraw/domain/account/application/port/input/MyFilmographyUseCase.kt
@@ -1,9 +1,0 @@
-package com.project.indistraw.domain.account.application.port.input
-
-import com.project.indistraw.domain.account.application.port.input.dto.FilmographyDto
-
-interface MyFilmographyUseCase {
-
-    fun execute(): List<FilmographyDto>
-
-}
