+++ src/main/kotlin/com/project/indistraw/domain/movie/application/port/input/MyFilmographyUseCase.kt
@@ -1,0 +1,9 @@
+package com.project.indistraw.domain.movie.application.port.input
+
+import com.project.indistraw.domain.movie.application.port.input.dto.FilmographyDto
+
+interface MyFilmographyUseCase {
+
+    fun execute(): List<FilmographyDto>
+
+}
