+++ src/main/kotlin/com/project/indistraw/domain/movie/application/port/input/dto/FilmographyDto.kt
@@ -1,4 +1,4 @@
-package com.project.indistraw.domain.account.application.port.input.dto
+package com.project.indistraw.domain.movie.application.port.input.dto
 
 data class FilmographyDto(
     val idx: Long,
